refactor: vendor unique name schema, preview, add is_deleted unit schema; fix: suspended_reason, suspended_until suspended_at for update, throw FK exeption unit delete

+++ src/models/vendors/delete-vendor/delete-vendor.service.ts
@@ -1,11 +1,14 @@
 import { Injectable } from '@nestjs/common';
 import { VendorRepository } from '../repositories';
+import { VendorNotFoundException } from '../exceptions';
 
 @Injectable()
 export class DeleteVendorService {
-  constructor(private readonly vendorRepository: VendorRepository) {}
+  constructor(private readonly vendorRepository: VendorRepository) { }
 
   async delete(id: string): Promise<void> {
+    const vendor = await this.vendorRepository.findById(id);
+    if (!vendor) throw new VendorNotFoundException(id);
     return this.vendorRepository.delete(id);
   }
 }
